# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Solvit API",
    "description": "Solvit — Next.js / PostgreSQL project by Femiznet. Test with the Scalar UI at /reference.",
    "version": "0.1.0",
    "contact": {
      "name": "Webhooks / feedback"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin API (works on localhost and preview deploys)"
    }
  ],
...
  "security": [
    {
      "cookieAuth": []
    },
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/users/{id}/likes": {
      "parameters": [
        {
          "$ref": "#/components/parameters/IdParam"
        }
      ],
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get a user's likes",
        "description": "Authenticated. Returns the projects and solutions liked by the requested user id.",
        "operationId": "getUserLikes",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Likes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "oneOf": [
                          {
                            "$ref": "#/components/schemas/Project"
                          },
                          {
                            "$ref": "#/components/schemas/Solution"
                          }
                        ]
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing user id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "solvit_session",
        "description": "HTTP-only session cookie set on signup/login. Same-origin browsers send it automatically."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT fallback returned in the body of /api/auth/signup and /api/auth/login."
      }
    },
    "parameters": {
      "IdParam": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        },
        "description": "UUID identifier"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string"
          },
          "fieldErrors": {
            "type": "object",
            "description": "Per-field validation errors when applicable."
          }
        }
      },
...
      "Project": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "instructions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "level": {
            "type": "string",
            "enum": [
              "BEGINNER",
              "INTERMEDIATE",
              "ADVANCED"
            ]
          },
          "requirements": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "optRequirements": {
            "oneOf": [
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              {
                "type": "null"
              }
            ]
          },
          "categoryId": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "totalLikes": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Solution": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "projectId": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "repoUrl": {
            "oneOf": [
              {
                "type": "string",
                "format": "uri"
              },
              {
                "type": "null"
              }
            ]
          },
          "demoUrl": {
            "oneOf": [
              {
                "type": "string",
                "format": "uri"
              },
              {
                "type": "null"
              }
            ]
          },
          "implFeat": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "totalLikes": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}